Change Symbol class attr and scanner reading behaviour; minor bug fixes
Instead of storing the same line multiple times for each symbol,
it is changed so that the scanner instance stores an array of lines,
which can be accessed by the line numbers.

The above lead to a change of behaviour in get_next_char(), which now
simply reads the array of lines and updates array indices.

This code has been tested on the two example files (First report/examples/
*.circuit) and found to produce correct results. Testing for edge cases,
as well as writing unit tests, are yet to be done.