Unit tests for Parser (syntax checks ONLY); bug fixes in Parser
Unit tests were written for the syntax-check-only parser (i.e.
the parser that is handed over to sa787 to add semantic checks).
In doing so, a few bugs were found and fixed. These changes should
be merged before checking semantics.

Note also that error handling (message output to user + recovery)
are also to be implemented by sa787, so the current tests only
tests for boolean return values indicating correctness, but does
not check whether the "correct" type of error is returned.