Added function for GUI to request logsim for an editor and a restart
A function, to be used for as a button on-click event handler, has
been added to gui.py. When called, this function terminates the GUI,
sets a flag for the request, and returns execution to logsim.py.
logsim.py would then check for the existence of this flag, and if
found, opens an editor, waits for the editor to terminate, and then
restart the GUI.