Compatibility fix for launching editor
Apparently, `editor' a.k.a. `/usr/bin/editor' is only a Debian
feature, and not guaranteed to exist on other *nix. (Most
importantly, does not exist on the CUED Teaching System.) As
such, the code is now modified such that it will attempt to
determine a user-specified editor through environment variables
VISUAL and EDITOR, then try a list of commonly available editors,
falling back to `vi' and `ed' if necessary (these are very likely
to exist by POSIX requirements). If even `ed' is not found (or
not in $PATH), then it will prompt the user to enter the name
(in $PATH), or full path, of an editor, or allow the user to
terminate the program.

It is noted that as a result of this change, the program can
now run arbitrary code as the same user who started it. The
justification is that the user who started this program would
be able to run other scripts anyway, but if this will become
untrue in the future then this behaviour MUST BE CHANGED.