Updated test_parse.py; bug fixes in various place
test_parse.py was updated to reflect the changed behaviour of
Parser, and also to pass actual instances of Network, Monitors
and Devices instead of placeholders None.

In doing so, a minor bug in parse.py has been discovered and fixed.

In gui.py, the previous commit allowed the position of the text
message to be fixed on screen with canvas panning, but not with
zooming. This is fixed now so that its position and size remains
fixed even the canvas zoom is changed.

In logsim.py it was discovered that Network, Devices, etc were
not re-initialised, leading to the new GUI instance retaining
old values. This has now been fixed.