Fixed bug in Scanner on multiline comments; added examples
It was discovered that when there are multiline unindented
comments, the continuation lines will not be skipped. This
commit fixes the problem, and added relevant test cases in
test_scanner.py.

Two example circuit definition files were added, an 8-to-1
multiplexer and a sequence detector. These can be used in
the 2nd interim report. The examples written for the 1st
interim report were also copied and placed in the `examples'
folder.