Devices: Implemented RC device
Maintenance work describes an RC device which has now been
implemented. This required changes in devices.py, and also
network.py which specifies how the device behaves.

It is noted that since the device has an initial state
memory, it needs to be reset when the network restarts.
A function for that purpose is created in devices.py, and
the GUI has been updated such that the function is called
when the "Restart" button is pressed.

Also, a minor bug has been fixed in logsim.py.